removed unsupported features

+++ tests/chapter_18/valid/no_structure_parameters/libraries/global_struct.h
@@ -1,7 +1,7 @@
 struct s {
     int i;
     char arr[2];
-    double d;
+    int d;
 };
 
 struct outer {
